Add a notice
A notice of how to enable the optional features.
I used to be annoyed by this problem for a long time because I tried to test again and again by different options of php, nginx and also h5ai itself. But I still cannot download packaged folders by "shell zip". The reason I choose h5ai is not only the beautiful interface but also the powerful functions, especially the multi-file download function.
Finally, I find out the right options but the process really makes me upset and tired. So I want to add this to notice others to let them be able to enable these functions directly.