Allow Custom Thumbnail
Add the ability to manually upload a thumbnail for a file. If there is a jpg with the same name in the subdirectory ".thumb", this will be rendered as thumbnail.
Good for e.g. video preview, if you don't have ffmpeg, or any other file where creation of thumbnail isn't supported.