Add thumbnail support for archive files
* Separate 7z from regular zip archive because PHP ZipArchive doesn't
handle them.
* Support RAR archives with PHP "rar" extension.
* Use sqlite3 database for caching errors from failures to process archive files.
* Add blocklist for more explicit filtering.
When a value is removed from a default array of strings in the
options, the removed value should be considered explicitly blocked by
the user: thumbnails should not be generated, actual underlying file type should
not be checked for. Thus the blocklist should be updated with the
missing string.
* Notify client if detected type is incorrect after checking the
database.
Check in database if we have cached the result of our server-side file
detection, return the cached file type if present.
* Return to client the same kind of type it uses in order to update the
previewer if the file type was wrongly detected.
* Remove client-side sample request for img as it is both superfulous and a security issue.
Down-sampling should be requested by the backend only. This can be implemented again on the
back-end if necessary.
* Fix double previewer bug:
Remove any previous callback on click event, as both previous and new previewers were triggered successively.
* Remove swf and flv file from handled video preview since these formats
are deprecated and should be handled by the browser directly.
* Display detected type in info panel.